Initial Commit
I am currently learning python. This is a small play project where am building a CLI to manage the books I am reading. This is the first small step. I have a menu where I can add, list, delete, or mark as read books from a python list. It is volatile so nothing will be persisted. The next thing to do will be to store these books in a file. I am not yet sure if I will go with CSV or JSON for this.